create corrections

+++ src/models/one_rep_max.py
@@ -3,6 +3,7 @@ from sqlmodel import SQLModel, Field
 
 from typing import Optional
 from datetime import datetime
+from decimal import Decimal
 
 
 class OneRepMax(SQLModel, table=True):
@@ -11,6 +12,6 @@ class OneRepMax(SQLModel, table=True):
     id: Optional[uuid.UUID] = Field(default_factory=uuid.uuid4, primary_key=True)
     user_id: uuid.UUID = Field(foreign_key="users.id", nullable=False, index=True)
     exercise_id: int = Field(foreign_key="exercises.id", nullable=False, index=True)
-    weight: float = Field(nullable=False)
+    weight: Decimal = Field(nullable=False)
     video_url: Optional[str] = Field(default=None)
     created_at: datetime = Field(default_factory=datetime.utcnow, nullable=False)
